Hide internal web runtime helpers
