+++ package.json
@@ -71,7 +71,7 @@
     "smoke:web-browser": "node scripts/smoke-web-browser.mjs",
     "smoke:cloudflare-worker": "node scripts/smoke-cloudflare-worker.mjs",
     "testcase:cloudflare-worker:prepare": "mkdir -p fixtures/manual/cloudflare-worker-issue-18/public/search && cargo run -p indexbind-build -- build-bundle fixtures/benchmark/basic/docs fixtures/manual/cloudflare-worker-issue-18/public/search/index.bundle hashing",
-    "testcase:cloudflare-worker:dev": "npm run testcase:cloudflare-worker:prepare && wrangler dev --config fixtures/manual/cloudflare-worker-issue-18/wrangler.jsonc",
+    "testcase:cloudflare-worker:dev": "npm run build && npm run testcase:cloudflare-worker:prepare && wrangler dev --config fixtures/manual/cloudflare-worker-issue-18/wrangler.jsonc",
     "check": "tsc --noEmit -p tsconfig.json && cargo check --workspace",
     "test": "cargo test --workspace",
     "release:verify": "node scripts/verify-release.mjs",
