+++ src/cloudflare.ts
@@ -1,6 +1,6 @@
 import wasmModule from './wasm/indexbind_wasm_bg.wasm';
 import { initSync, WasmIndex } from './wasm/indexbind_wasm.js';
-import { openWebIndexWithBindings } from './web.js';
+import { openWebIndexWithBindings } from './web-core.js';
 
 export {
   WebIndex,
@@ -11,11 +11,14 @@ export {
   type RerankerOptions,
   type SearchOptions,
   type WebArtifactInfo,
-} from './web.js';
+} from './web-core.js';
 
 let wasmInitialized = false;
 
-export async function openWebIndex(base: string | URL, options = {}) {
+export async function openWebIndex(
+  base: string | URL,
+  options: import('./web-core.js').OpenWebIndexOptions = {},
+) {
   if (!wasmInitialized) {
     initSync({ module: wasmModule });
     wasmInitialized = true;
